+++ banco-digital/padrao-camadas/src/main/java/br/com/ewerton/padraocamadas/domain/PessoaFisica.java
@@ -4,6 +4,8 @@ import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "tb_pessoa_fisica")
 public class PessoaFisica extends Pessoa {
@@ -18,7 +20,7 @@ public class PessoaFisica extends Pessoa {
     @Pattern(regexp = "^[0-9]{3}\\.[0-9]{3}\\.[0-9]{3}\\-[0-9]{2}$", message = "CPF deve estar no formato XXX.XXX.XXX-XX")
     private String pessoaFisicaCpf;
 
-    public PessoaFisica(Long pessoaFisicaID, String pessoaNome, String pessoaEmail, Double pessoaSaldo, Integer pessoaSenha, String pessoaFisicaCpf) {
+    public PessoaFisica(Long pessoaFisicaID, String pessoaNome, String pessoaEmail, BigDecimal pessoaSaldo, Integer pessoaSenha, String pessoaFisicaCpf) {
         super(pessoaNome, pessoaEmail, pessoaSaldo, pessoaSenha);
         this.pessoaFisicaCpf = pessoaFisicaCpf;
         this.pessoaFisicaID = pessoaFisicaID;
